Format and extend README.md
This also needs an explanation of how to build the front end and connect to the back end. Just putting the server URL in the browser gives a 404. I took a shot at building the front end with `npm install` and `npm run build`, but index.html just gives me a blank page and I wasn't sufficiently motivated to get into your code to get it working ;-).